tweak: get kiosk menu before menu page
