+++ store/src/sections/kiosk/index.js
@@ -16,6 +16,7 @@ import { KioskCart } from '../../components/kiosk/cart'
 import { DELETE_CART } from '../../graphql'
 import { usePayment } from '../../lib'
 import { InfoBar } from '../../components/kiosk/component'
+import { useKioskMenu } from './utils/useKioskMenu'
 // idle screen component
 // fulfillment component
 // header
@@ -42,6 +43,14 @@ const Kiosk = props => {
    // console.log('kioskConfig', kioskConfig)
 
    const { direction } = useTranslation()
+   const collectionIds = React.useMemo(() => {
+      return kioskConfig.data.collectionData.value.length > 0
+         ? kioskConfig.data.collectionData.value.map(
+              collection => collection.id
+           )
+         : []
+   }, [kioskConfig])
+   const { status: kioskMenuStatus, hydratedMenu } = useKioskMenu(collectionIds)
 
    const { resetPaymentProviderStates } = usePayment()
 
@@ -172,6 +181,8 @@ const Kiosk = props => {
                         config={kioskConfig}
                         setCurrentPage={setCurrentPage}
                         combinedCartItems={combinedCartItems}
+                        hydratedMenu={hydratedMenu}
+                        status={kioskMenuStatus}
                      />
                   </Content>
                )}
